Add: question views field when querying question

+++ src/main/java/com/chwihae/domain/comment/CommentRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 @Repository
 public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 
-    long countByQuestionEntityId(Long questionId);
+    int countByQuestionEntityId(Long questionId);
 
     long countByUserEntityId(Long userId);
