+++ src/main/java/com/chwihae/domain/vote/VoteRepository.java
@@ -17,7 +17,7 @@ public interface VoteRepository extends JpaRepository<VoteEntity, Long> {
 
     long countByUserEntityId(Long userId);
 
-    long countByQuestionEntityId(Long questionId);
+    int countByQuestionEntityId(Long questionId);
 
     Optional<VoteEntity> findByQuestionEntityIdAndUserEntityId(Long questionId, Long userId);
